+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   * Added support for Domain Validation endpoints:
     * [AddDomains](https://techdocs.akamai.com/domain-validation/reference/post-domains) - Adds domains to validate.
     * [DeleteDomain](https://techdocs.akamai.com/domain-validation/reference/delete-domain) - Deletes a domain validation.
+    * [DeleteDomains](https://techdocs.akamai.com/domain-validation/reference/delete-domains) - Deletes a batch of domains validation.
     * [InvalidateDomain](https://techdocs.akamai.com/domain-validation/reference/post-invalidate-domain) - Invalidates the specified domain.
     * [InvalidateDomains](https://techdocs.akamai.com/domain-validation/reference/post-invalidate-domains) - Invalidates the specified domains.
     * [ValidateDomains](https://techdocs.akamai.com/domain-validation/reference/post-validate-domains) - Immediately validates the domains using the DNS CNAME, DNS TXT, and HTTP methods.